+++ Makefile
@@ -195,8 +195,8 @@ test-release: Dockerfile_build
 		./gotool.sh github.com/goreleaser/goreleaser release --rm-dist --snapshot"
 
 
-bin/flux: $(STDLIB_SOURCES) $$(call go_deps,./internal/cmd/flux)
-	$(GO_BUILD) -o ./bin/flux ./internal/cmd/flux
+bin/flux: $(STDLIB_SOURCES) $$(call go_deps,./cmd/flux)
+	$(GO_BUILD) -o ./bin/flux ./cmd/flux
 
 
 libflux/target/release/fluxc: libflux
